fix: refactor backend packages
