+++ BOBA/BOBA.Server/Controllers/UserController.cs
@@ -1,4 +1,5 @@
 ﻿using BOBA.Server.Data;
+using BOBA.Server.Data.implementation;
 using BOBA.Server.Models.Dto;
 using BOBA.Server.Services.Interfaces;
 using Microsoft.AspNetCore.Authorization;
