+++ BOBA/BOBA.Server/Data/ApplicationDbContext.cs
@@ -1,6 +1,9 @@
 ﻿using BOBA.Server.Data.EntityTypeConfigurations;
+using BOBA.Server.Data.implementation;
+using BOBA.Server.Data.model;
 using Microsoft.AspNetCore.Identity.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore;
+using Task = BOBA.Server.Data.implementation.Task;
 
 namespace BOBA.Server.Data;
