+++ BOBA/BOBA.Server/Data/EntityTypeConfigurations/WorkflowSeedConfig.cs
@@ -1,4 +1,5 @@
-﻿using Microsoft.EntityFrameworkCore;
+﻿using BOBA.Server.Data.model;
+using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Metadata.Builders;
 
 namespace BOBA.Server.Data.EntityTypeConfigurations;
