+++ BOBA/BOBA.Server/Data/model/Choice.cs
@@ -1,8 +1,8 @@
 ﻿using System.ComponentModel.DataAnnotations;
 
-namespace BOBA.Server.Data;
+namespace BOBA.Server.Data.model;
 
-public class TaskField
+public class Choice
 {
     [Key]
     public string Id { get; set; }
@@ -13,10 +13,4 @@ public class TaskField
 
     [StringLength(500)]
     public string Description { get; set; }
-
-    [StringLength(100)]
-    public string? Validation { get; set; }
-
-    [StringLength(1000)]
-    public string? Options { get; set; }
 }
