+++ BOBA/BOBA.Server/Data/model/TaskState.cs
@@ -1,7 +1,9 @@
-﻿using System.ComponentModel.DataAnnotations;
+﻿using BOBA.Server.Data.implementation;
+using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
+using Task = BOBA.Server.Data.implementation.Task;
 
-namespace BOBA.Server.Data
+namespace BOBA.Server.Data.model
 {
     public class TaskState
     {
