+++ BOBA/BOBA.Server/Services/UserService.cs
@@ -1,4 +1,6 @@
 ﻿using BOBA.Server.Data;
+using BOBA.Server.Data.implementation;
+using BOBA.Server.Data.model;
 using BOBA.Server.Models.Dto;
 using BOBA.Server.Services.Interfaces;
 using Microsoft.AspNetCore.Identity;
